requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.7.0
anthropic==0.28.0
anyio==4.4.0
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
asttokens==2.4.1
attrs==23.2.0
Babel==2.15.0
backoff==2.2.1
bcrypt==4.1.3
black==23.3.0
blinker==1.8.2
build==1.2.1
CacheControl==0.14.0
cachetools==5.3.3
certifi==2024.6.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.5.3
cleo==2.1.0
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
coverage==7.5.3
crashtest==0.4.1
cryptography==43.0.0
cssselect2==0.7.0
dataclasses-json==0.5.7
datasets==2.20.0
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
deprecation==2.1.0
dill==0.3.8
distlib==0.3.8
distro==1.9.0
dj-database-url==1.3.0
Django==4.2.13
django-widget-tweaks==1.5.0
dnspython==2.6.1
dulwich==0.21.7
email_validator==2.2.0
executing==2.0.1
faiss-cpu==1.8.0.post1
fastapi==0.111.0
fastapi-cli==0.0.4
fastembed==0.3.1
fastjsonschema==2.19.1
filelock==3.15.1
findspark==2.0.1
flatbuffers==24.3.25
forge==0.22.0
forge-core==0.4.0
forge-db==0.3.4
forge-heroku==0.4.0
forge-tailwind==0.3.4
forge-work==0.6.0
frozenlist==1.4.1
fsspec==2024.5.0
ghp-import==2.1.0
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.6
google-api-core==2.19.0
google-api-python-client==2.132.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.2
googleapis-common-protos==1.63.1
gpt-engineer==0.3.1
greenlet==3.0.3
grpcio==1.64.1
grpcio-status==1.62.2
gunicorn==20.1.0
h11==0.14.0
hiredis==2.3.2
honcho==1.1.0
html5lib==1.1
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.3
humanfriendly==10.0
hupper==1.12.1
idna==3.7
importlib_metadata==7.1.0
importlib_resources==6.4.0
iniconfig==2.0.0
installer==0.7.0
intel-openmp==2021.4.0
ipykernel==6.29.4
ipython==8.23.0
jaraco.classes==3.4.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.4.2
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.1
jupyter_core==5.7.2
keyring==24.3.1
kubernetes==30.1.0
langchain==0.2.3
langchain-anthropic==0.1.15
langchain-community==0.2.4
langchain-core==0.2.22
langchain-google-genai==1.0.7
langchain-openai==0.1.8
langchain-text-splitters==0.2.1
langsmith==0.1.77
loguru==0.7.2
lxml==5.2.2
Markdown==3.6
markdown-it-py==3.0.0
markdown2==2.5.0
MarkupSafe==2.1.5
marshmallow==3.21.3
marshmallow-enum==1.5.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mergedeep==1.3.4
mkdocs==1.6.0
mkdocs-get-deps==0.2.0
mkdocs-material==9.5.26
mkdocs-material-extensions==1.3.1
mkdocs-table-reader-plugin==2.2.2
mkl==2021.4.0
mmh3==4.1.0
monotonic==1.6
more-itertools==10.3.0
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
multiprocess==0.70.16
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
numpy==1.26.4
oauthlib==3.2.2
onnx==1.16.1
onnxruntime==1.18.1
openai==1.34.0
opentelemetry-api==1.25.0
opentelemetry-exporter-otlp-proto-common==1.25.0
opentelemetry-exporter-otlp-proto-grpc==1.25.0
opentelemetry-instrumentation==0.46b0
opentelemetry-instrumentation-asgi==0.46b0
opentelemetry-instrumentation-fastapi==0.46b0
opentelemetry-proto==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-semantic-conventions==0.46b0
opentelemetry-util-http==0.46b0
orjson==3.10.4
oscrypto==1.3.0
overrides==7.7.0
packaging==23.2
paginate==0.5.6
pandas==2.2.2
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pillow==10.3.0
pkginfo==1.11.1
platformdirs==4.2.0
pluggy==1.5.0
poetry==1.8.3
poetry-core==1.9.0
poetry-plugin-export==1.8.0
posthog==3.5.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.8
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
py4j==0.10.9.7
pyarrow==16.1.0
pyarrow-hotfix==0.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.7.3
pydantic_core==2.18.4
pydeck==0.9.1
Pygments==2.17.2
pyHanko==0.25.1
pyhanko-certvalidator==0.26.3
pymdown-extensions==10.8.1
PyMuPDF==1.24.7
PyMuPDFb==1.24.6
pyparsing==3.1.2
pypdf==4.3.1
PyPDF2==3.0.1
pyperclip==1.8.2
PyPika==0.48.9
pypng==0.20220715.0
pyproject_hooks==1.1.0
pyreadline3==3.4.1
pyspark==3.5.1
PyStemmer==2.2.0.1
pytest==7.4.4
pytest-django==4.8.0
python-bidi==0.5.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
pywin32==306
pywin32-ctypes==0.2.2
PyYAML==6.0.1
pyyaml_env_tag==0.1
pyzmq==25.1.2
qrcode==7.4.2
rapidfuzz==3.9.3
redis==4.6.0
referencing==0.35.1
regex==2023.12.25
reportlab==4.2.2
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.7.1
rpds-py==0.18.1
rsa==4.9
rudder-sdk-python==2.1.1
ruff==0.0.194
safetensors==0.4.3
scikit-learn==1.5.1
scipy==1.14.0
sentence-transformers==3.0.1
setuptools==70.0.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
snowballstemmer==2.2.0
SQLAlchemy==2.0.30
sqlparse==0.5.0
stack-data==0.6.3
starlette==0.37.2
streamlit==1.35.0
svglib==1.5.1
sympy==1.12.1
tabulate==0.9.0
tbb==2021.13.0
tenacity==8.3.0
termcolor==2.3.0
threadpoolctl==3.5.0
tiktoken==0.7.0
tinycss2==1.3.0
tokenizers==0.19.1
toml==0.10.2
tomlkit==0.12.5
toolz==0.12.1
torch==2.3.1
tornado==6.4
tqdm==4.66.4
traitlets==5.14.2
transformers==4.42.3
trove-classifiers==2024.5.22
typer==0.12.3
typing-inspect==0.9.0
typing_extensions==4.12.1
tzdata==2024.1
tzlocal==5.2
ujson==5.10.0
uritemplate==4.1.1
uritools==4.0.3
urllib3==2.2.1
uvicorn==0.30.1
virtualenv==20.26.2
watchdog==4.0.1
watchfiles==0.22.0
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
websockets==12.0
whitenoise==6.6.0
win32-setctime==1.1.0
wrapt==1.16.0
xhtml2pdf==0.2.16
xxhash==3.4.1
yarl==1.9.4
zipp==3.19.2
